Average parameter estimates across replicates, and regularize variance estimates

AverageAndRegularize uses a linear model to average estimates of a parameter of interest over replicates, and to get averages for all of a set of conditions specified by the user. This specification is done through formulas that will be used to create a design matrix for parameters to estimate. Currently, some design matrices will yield much longer runtimes than others. This is due to some design matrices necessitating numerically estimating a set of maximum likelihood parameters, while others yield simple, analytic approximations that are quick to calculate. For example, formulas such as ~ treatment or ~ treatment:duration will be quick to fit, and formulas such as ~ treatment + batch are currently much slower to fit. This will eventually be remedied through approximation of the solution to all full models.

Usage

AverageAndRegularize(
  obj,
  features = NULL,
  parameter = "log_kdeg",
  type = "kinetics",
  kstrat = NULL,
  populations = NULL,
  fraction_design = NULL,
  exactMatch = TRUE,
  repeatID = NULL,
  formula_mean = NULL,
  formula_sd = NULL,
  include_all_parameters = TRUE,
  sd_reg_factor = 10,
  error_if_singular = TRUE,
  min_reads = 10,
  force_optim = FALSE,
  conservative = FALSE,
  character_limit = 20,
  overwrite = TRUE
)

Arguments

obj

An EZbakRFractions object, which is an EZbakRData object on which EstimateFractions() has been run.

features

Character vector of the set of features you want to stratify reads by and estimate proportions of each RNA population. The default of "all" will use all feature columns in the obj's cB.

parameter

Parameter to average across replicates of a given condition.

type

What type of table is the parameter found in? Default is "kinetics", but can also set to "fractions".

kstrat

If type == "kinetics", then kstrat specifies the kinetic parameter inference strategy.

exactMatch

If TRUE, then features and populations have to exactly match those for a given fractions table for that table to be used. Means that you can't specify a subset of features or populations by default, since this is TRUE by default.

repeatID

If multiple kinetics or fractions tables exist with the same metadata, then this is the numerical index by which they are distinguished.

formula_mean

An R formula object specifying how the parameter of interest depends on the sample characteristics specified in obj's metadf. The most common formula will be ~ treatment or ~ treatment:duration, where treatment and duration would be replaced with whatever you called the relevant sample characteristics in your metadf. ~ treatment means that an average value of parameter should be estimated for each set of samples with the same value for treatment in the metadf. ~ treatment:duration specifies that an average value of parameter should be estimated for each set of samples with the same combination of treatment and duration values in the metadf. An example of the latter case is a situation where you have two or more treatments (e.g., drug treated and untreated control) which were applied for different durations of time (e.g., 4 and 8 hours).

formula_sd

Same as formula_mean, but this time specifying how the variance in the replicate estimates of parameter depends on the sample characteristics specified in obj's metadf. Unlike standard linear modeling, this allows you to specify a heteroskedastic model. I suggest allowing a parameter's variance to depend on the "treatment" condition, as changes in relative RNA abundance can impact parameter variance, so differential expression caused by your "treatment" could impact parameter variance.

include_all_parameters

If TRUE, an additional table will be saved with the prefix fullfit_, which includes all of the parameters estimated throughout the course of linear modeling and regularization. This can be nice for visualizing the regularized mean-variance trend.

sd_reg_factor

Determines how strongly variance estimates are shrunk towards trend. Higher numbers lead to more regularization. Eventually, this will be replaced with estimation of how much variance there seems to be in the population of variances.

error_if_singular

If TRUE, linear model will throw an error if parameters cannot be uniquely identified. This is most often caused by parameters that cannot be estimated from the data, e.g., due to limited replicate numbers or correlated sample characteristics (i.e., all treatment As also correspond to batch As, and all treatment Bs correspond to batch Bs).

min_reads

Minimum number of reads in all samples for a feature to be kept.

force_optim

Perform numerical likelihood estimation, even if a more efficient, approximate, analytical strategy is possible given formula_mean and formula_sd.

conservative

If TRUE, conservative variance regularation will be performed. In this case, variances below the trend will be regularized up to the trend, and variances above the trend will be left unregularized. This avoids underestimation of variances.

character_limit

Limit on the number of characters of the name given to the output table. Will attempt to concatenate the parameter name with the names of all of the features. If this is too long, only the parameter name will be used.

overwrite

If TRUE, identical, existing output will be overwritten.