Make an MAPlot from EZbakR comparison

Make an MAPlot from EZbakR comparison

Usage

EZMAPlot(
  obj,
  parameter = "log_kdeg",
  design_factor = NULL,
  reference = NULL,
  experimental = NULL,
  param_name = NULL,
  param_function = NULL,
  features = NULL,
  condition = NULL,
  repeatID = NULL,
  exactMatch = TRUE,
  plotlog2 = TRUE,
  FDR_cutoff = 0.05,
  difference_cutoff = log(2),
  size = NULL
)

Arguments

obj

An object of class EZbakRCompare, which is an EZbakRData object on which you have run CompareParameters

parameter

Name of parameter whose comparison you want to plot.

design_factor

Name of factor from metadf whose parameter estimates at different factor values you would like to compare.

reference

Name of reference condition factor level value.

experimental

Name of condition factor level value to compare to reference.

param_name

If you want to assess the significance of a single parameter, rather than the comparison of two parameters, specify that one parameter's name here.

param_function

NOT YET IMPLEMENTED. Will allow you to specify more complicated functions of parameters when hypotheses you need to test are combinations of parameters rather than individual parameters or simple differences in two parameters.

features

Character vector of feature names for which comparisons were made.

condition

Defunct parameter that has been replaced with design_factor. If provided gets passed to design_factor if design_factor is not already specified.

repeatID

If multiple kinetics or fractions tables exist with the same metadata, then this is the numerical index by which they are distinguished.

exactMatch

If TRUE, then features and populations have to exactly match those for a given fractions table for that table to be used. Means that you can't specify a subset of features or populations by default, since this is TRUE by default.

plotlog2

If TRUE, assume that log(parameter) difference is passed in and that you want to plot log2(parameter) difference.

FDR_cutoff

False discovery cutoff by which to color points.

difference_cutoff

Minimum absolute difference cutoff by which to color points.

size

Size of points, passed to geom_point() size parameter. If not specified, a point size is automatically chosen.