Make a VolcanoPlot from EZbakR comparison

Make a VolcanoPlot from EZbakR comparison

Usage

EZVolcanoPlot(
  obj,
  parameter = "log_kdeg",
  condition = NULL,
  reference = NULL,
  experimental = NULL,
  features = NULL,
  exactMatch = TRUE,
  plotlog2 = TRUE,
  FDR_cutoff = 0.05,
  difference_cutoff = 0,
  size = NULL
)

Arguments

obj

An object of class EZbakRCompare, which is an EZbakRData object on which you have run CompareParameters

parameter

Name of parameter whose comparison you want to plot.

condition

Name of condition for which you want to plot the comparison of two of its levels. Condition should be the name of a column that appears in your metadf table.

reference

Name of reference level for comparison.

experimental

Name of experimental level for comparison.

features

Character vector of feature names for which comparisons were made.

exactMatch

If TRUE, then features has to exactly match those for a given comparisons table for that table to be used. Means that you can't specify a subset of features by default, since this is TRUE by default.

plotlog2

If TRUE, assume that log(parameter) difference is passed in and that you want to plot log2(parameter) difference. TO-DO: probably best to change this to a more general scale parameter by which the parameter is multiplied. Default would be log2(exp(1)) to convert log() to log2().

FDR_cutoff

False discovery cutoff by which to color points.

difference_cutoff

Minimum absolute difference cutoff by which to color points.

size

Size of points, passed to geom_point() size parameter. If not specified, a point size is automatically chosen.