Easily get EZbakR table of estimates of interest

Easily get EZbakR table of estimates of interest

Usage

EZget(
  obj,
  type = c("fractions", "kinetics", "readcounts", "averages", "comparisons"),
  features = NULL,
  populations = NULL,
  fraction_design = NULL,
  isoforms = NULL,
  kstrat = NULL,
  parameter = NULL,
  counttype = NULL,
  condition = NULL,
  experimental = NULL,
  reference = NULL,
  repeatID = NULL,
  returnNameOnly = FALSE,
  exactMatch = FALSE,
  alwaysCheck = FALSE
)

Arguments

obj

EZbakRData object

type

The class of EZbakR outputs would you like to search through. Equivalent to the name of the list in the EZbakRData object that contains the tables of interest.

features

Features that must be present in the table of interest. If exactMatch is TRUE, then these features must also be the only features present in the table.

populations

Only relevant if type == "fractions". Mutational populations that must have been analyzed to generate the table of interest.

fraction_design

Only relevant if type == "fractions". Fraction design table used to generate the table of interest.

isoforms

If the relevant table is the result of isoform deconvolution

kstrat

Only relevant if type == "kinetics". Short for "kinetics strategy"; the strategy used to infer kinetic parameters.

parameter

Only relevant if type == "averages" or "comparisons". Which parameter was being averaged or compared?

counttype

String denoting what type of read count information you are looking for. Current options are "TMM_normalized", "transcript", and "matrix". TO-DO: Not sure this is being used in any way currently...

condition

Condition specified in relevant run of CompareParameters(). Therefore, only relevant if type == "comparisons".

experimental

Experimental condition specified in relevant run of CompareParameters(). Therefore, only relevant if type == "comparisons".

reference

Reference condition specified in relevant run of CompareParameters(). Therefore, only relevant if type == "comparisons".

repeatID

Numerical ID for duplicate objects with same metadata.

returnNameOnly

If TRUE, then only the names of tables that passed your search criteria will be returned. Else, the single table passing your search criteria will be returned. If there is more than one table that passes your search criteria and returnNameOnly == FALSE, an error will be thrown.

exactMatch

If TRUE, then for features and populations, which can be vectors, ensure that provided vectors of features and populations exactly match the relevant metadata vectors.

alwaysCheck

If TRUE, then even if there is only a single table for the type of interest, still run all checks against queries.