Import transcript isoform quantification into EZbakRData object

A convenient wrapper to tximport() for importing isoform quantification data into an EZbakRData object. This is necessary to run functions such as EstimateIsoformFractions.

Usage

ImportIsoformQuant(
  obj,
  files,
  quant_tool = c("none", "salmon", "sailfish", "alevin", "piscem", "kallisto", "rsem",
    "stringtie"),
  ...
)

Arguments

obj

An EZbakRData object.

files

A named vector of paths to all transcript quantification files that you would like to import. This will be passed as the first argument of tximport::tximport() (also named files). The names of this vector should be the same as the sample names as they appear in the metadf of the EZbakRData object.

quant_tool

String denoting the type of software used to generate the abundances. Will get passed to the type argument of tximport::tximport(). As described in the documentation for tximport 'Options are "salmon", "sailfish", "alevin", "piscem", "kallisto", "rsem", "stringtie", or "none". This argument is used to autofill the arguments below (geneIdCol, etc.) "none" means that the user will specify these columns. Be aware that specifying type other than "none" will ignore the arguments below (geneIdCol, etc.)'. Referenced 'arguments below' can be specified as part of ....

...

Additional arguments to be passed to tximport::tximport(). Especially relevant if you set quant_tool to "none".