Get contrasts of estimated parameters

Get contrasts of estimated parameters

Usage

CompareParameters(
  obj,
  condition,
  reference,
  experimental,
  features = NULL,
  exactMatch = TRUE,
  repeatID = NULL,
  overwrite = TRUE,
  parameter = "log_kdeg"
)

Arguments

obj

An EZbakRFit object, which is an EZbakRFractions object on which AverageAndRegularize() has been run.

condition

Name of factor from metadf whose parameter estimates at different factor values you would like to compare.

reference

Name of reference condition factor level value. Difference will be calculated as experimental - reference.

experimental

Name of condition factor level value to compare to reference. Difference will be calculated as experimental - reference.

features

Character vector of the set of features you want to stratify reads by and estimate proportions of each RNA population. The default of "all" will use all feature columns in the obj's cB.

exactMatch

If TRUE, then features and populations have to exactly match those for a given fractions table for that table to be used. Means that you can't specify a subset of features or populations by default, since this is TRUE by default.

repeatID

If multiple averages tables exist with the same metadata, then this is the numerical index by which they are distinguished.

overwrite

If TRUE, then identical output will be overwritten if it exists.

parameter

Parameter to average across replicates of a given condition.