Generate a fraction_design table for EstimateFractions

Generate a fraction_design table for EstimateFractions

Usage

create_fraction_design(mutrate_populations)

Arguments

mutrate_populations

Character vector of the set of mutational populations present in your data. For example, s4U fed data with standard nucleotide recoding chemistry (e.g., TimeLapse, SLAM, TUC, AMUC, etc.) would have a mutrate_populations of c("TC"). Dual labeling experiments with s4U and s6G feeds would have a mutrate_populations of c("TC", "GA").

Value

A fraction_design table that assumes that every possible combination of mutational populations listed in mutrate_populations are present in your data. The present column can be modified if this assumption is incorrect.